requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
billiard==4.2.0
bleak==0.22.2
boto3==1.34.113
botocore==1.34.113
Brotli==1.1.0
carehare==1.0.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-rabbitmq==4.0.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.7
cssselect2==0.7.0
dbus-fast==2.21.3
distlib==0.3.8
Django==5.0.6
django-celery==3.1.17
django-celery-results==2.5.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-livesync==0.5
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-storages==1.14.3
django-tailwindcss==0.4.0
et-xmlfile==1.1.0
filelock==3.15.1
flower==2.0.1
fonttools==4.52.1
frozenlist==1.4.1
gunicorn==22.0.0
html5lib==1.1
humanize==4.9.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
loguru==0.7.2
lxml==5.2.1
msgpack==1.0.8
multidict==6.0.5
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.0
pamqp==3.3.0
phonenumbers==8.13.36
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pycups==2.0.4
pydyf==0.10.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tornado==6.4
twilio==9.1.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
whitenoise==6.6.0
xhtml2pdf==0.2.15
yarl==1.9.4
zopfli==0.2.3
